refactor code

+++ setup_ubuntu.sh
@@ -47,7 +47,7 @@ install_package() {
         stdbuf -oL sudo apt install -y "$1"
         # double check if the package was re-installed successfully
         if sudo dpkg -l | grep -q -w "$1"; then
-            printf "${OK} ${YELLOW}$1${RESET} has been installed successfully\n"
+            printf "${OK} ${SKY_BLUE}$1${RESET} has been installed successfully\n"
         else
             printf "${ERROR} ${YELLOW}$1${RESET} failed to install.\n"
         fi
@@ -57,13 +57,13 @@ install_package() {
 apt_update() {
     printf "${NOTE} Updating ${SKY_BLUE}apt${RESET} ...\n"
     sudo apt update -y
-    printf "${OK} apt has been updated successfully\n"
+    printf "${OK} ${SKY_BLUE}apt${RESET} has been updated successfully\n"
 }
 
 apt_upgrade() {
     printf "${NOTE} Upgrading ${SKY_BLUE}apt${RESET} ...\n"
     sudo apt upgrade -y
-    printf "${OK} apt has been upgraded successfully\n"
+    printf "${OK} ${SKY_BLUE}apt${RESET} has been upgraded successfully\n"
 }
 
 detect_os_info
@@ -81,7 +81,11 @@ printf "${BLUE}"===================================================="${RESET}\n"
 
 # ----------install some basic packages----------
 apt_update
+printf "\n%.0s" {1..1}
+
 apt_upgrade
+printf "\n%.0s" {1..1}
+
 printf "${NOTE} Installing ${SKY_BLUE}some basic packages${RESET} ...${RESET}\n"
 install_package "git"
 install_package "curl"
@@ -100,12 +104,11 @@ printf "${NOTE} Installing ${SKY_BLUE}core zsh${RESET} ...${RESET}\n"
 install_package "zsh"
 install_package "mercurial"
 install_package "zplug"
-
 printf "\n%.0s" {1..1}
 
 
 # ----------install oh-my-zsh, plugins----------
-if command -v zsh >/dev/null; then
+if command -v zsh >/dev/null 2>&1; then
     printf "${NOTE} Installing ${SKY_BLUE}Oh My Zsh and plugins${RESET} ...\n"
 
     if [ ! -d "$HOME/.oh-my-zsh" ]; then  
@@ -164,15 +167,15 @@ printf "\n%.0s" {1..1}
 
 
 # ----------install docker----------
-apt_update
 printf "${NOTE} Installing ${SKY_BLUE}docker, docker compose${RESET} ...${RESET}\n"
-if ! command -v docker >/dev/null; then
+apt_update
+if ! command -v docker >/dev/null 2>&1; then
     curl -fsSL https://get.docker.com -o get-docker.sh
     sh get-docker.sh
     rm get-docker.sh
-    printf "${OK} Docker has been installed successfully\n"
+    printf "${OK} ${SKY_BLUE}docker${RESET} has been installed successfully\n"
 else
-    printf "${OK} Docker is already installed. Skipping\n"
+    printf "${OK} ${MAGENTA}docker${RESET} is already installed. Skipping\n"
 fi
 apt_update
 install_package "docker-compose-plugin"
@@ -188,48 +191,68 @@ if ! command -v conda >/dev/null 2>&1; then
     $HOME/miniconda/bin/conda init bash
     $HOME/miniconda/bin/conda init zsh
     rm miniconda.sh
-    printf "${OK} Miniconda has been installed successfully\n"
+    printf "${OK} ${SKY_BLUE}miniconda${RESET} has been installed successfully\n"
 else
     $HOME/miniconda/bin/conda init bash
     $HOME/miniconda/bin/conda init zsh
-    printf "${OK} Miniconda is already installed. Skipping\n"
+    printf "${OK} ${MAGENTA}miniconda${RESET} is already installed. Skipping\n"
 fi
-printf "\n%.0s" {1..1}  
+printf "\n%.0s" {1..1}
 
 
 # ----------install vscode----------
 printf "${NOTE} Installing ${SKY_BLUE}Visual Studio Code${RESET} ...${RESET}\n"
-wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > packages.microsoft.gpg
-sudo install -D -o root -g root -m 644 packages.microsoft.gpg /etc/apt/keyrings/packages.microsoft.gpg
-echo "deb [arch=amd64,arm64,armhf signed-by=/etc/apt/keyrings/packages.microsoft.gpg] https://packages.microsoft.com/repos/code stable main" | sudo tee /etc/apt/sources.list.d/vscode.list > /dev/null
-rm -f packages.microsoft.gpg
-apt_update
-install_package "code"
+if ! command -v code >/dev/null 2>&1; then
+    wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor > packages.microsoft.gpg
+    sudo install -D -o root -g root -m 644 packages.microsoft.gpg /etc/apt/keyrings/packages.microsoft.gpg
+    echo "deb [arch=amd64,arm64,armhf signed-by=/etc/apt/keyrings/packages.microsoft.gpg] https://packages.microsoft.com/repos/code stable main" | sudo tee /etc/apt/sources.list.d/vscode.list > /dev/null
+    rm -f packages.microsoft.gpg
+    apt_update
+    install_package "code"
+else
+    printf "${OK} ${MAGENTA}Visual Studio Code${RESET} is already installed. Skipping\n"
+fi
+printf "\n%.0s" {1..1}
 
 
 # ----------install ms edge----------
 printf "${NOTE} Installing ${SKY_BLUE}Microsoft Edge${RESET} ...${RESET}\n"
-echo "deb [arch=amd64 signed-by=/etc/apt/keyrings/packages.microsoft.gpg] https://packages.microsoft.com/repos/edge stable main" | sudo tee /etc/apt/sources.list.d/microsoft-edge.list > /dev/null
-apt_update
-install_package "microsoft-edge-stable"
+if ! command -v microsoft-edge-stable >/dev/null 2>&1; then
+    echo "deb [arch=amd64 signed-by=/etc/apt/keyrings/packages.microsoft.gpg] https://packages.microsoft.com/repos/edge stable main" | sudo tee /etc/apt/sources.list.d/microsoft-edge.list > /dev/null
+    apt_update
+    install_package "microsoft-edge-stable"
+else
+    printf "${OK} ${MAGENTA}Microsoft Edge${RESET} is already installed. Skipping\n"
+fi
+printf "\n%.0s" {1..1}
 
 
 # ----------install spotify----------
 printf "${NOTE} Installing ${SKY_BLUE}Spotify${RESET} ...${RESET}\n"
-curl -sS https://download.spotify.com/debian/pubkey_C85668DF69375001.gpg | sudo gpg --dearmor --yes -o /etc/apt/trusted.gpg.d/spotify.gpg
-echo "deb http://repository.spotify.com stable non-free" | sudo tee /etc/apt/sources.list.d/spotify.list
-apt_update
-install_package "spotify-client"
+if ! command -v spotify >/dev/null 2>&1; then
+    curl -sS https://download.spotify.com/debian/pubkey_C85668DF69375001.gpg | sudo gpg --dearmor --yes -o /etc/apt/trusted.gpg.d/spotify.gpg
+    echo "deb http://repository.spotify.com stable non-free" | sudo tee /etc/apt/sources.list.d/spotify.list
+    apt_update
+    install_package "spotify-client"
+else
+    printf "${OK} ${MAGENTA}Spotify${RESET} is already installed. Skipping\n"
+fi
+printf "\n%.0s" {1..1}
 
 
 # ----------install ibus-bamboo----------
 printf "${NOTE} Installing ${SKY_BLUE}ibus-bamboo${RESET} ...${RESET}\n"
-sudo add-apt-repository -y ppa:bamboo-engine/ibus-bamboo
-apt_update
-install_package "ibus"
-install_package "ibus-bamboo"
-ibus restart
-env DCONF_PROFILE=ibus dconf write /desktop/ibus/general/preload-engines "['BambooUs', 'Bamboo']" && gsettings set org.gnome.desktop.input-sources sources "[('xkb', 'us'), ('ibus', 'Bamboo')]"
+if ! command -v ibus >/dev/null 2>&1; then
+    printf "${NOTE} Installing ${SKY_BLUE}ibus-bamboo${RESET} ...${RESET}\n"
+    sudo add-apt-repository -y ppa:bamboo-engine/ibus-bamboo
+    apt_update
+    install_package "ibus"
+    install_package "ibus-bamboo"
+    ibus restart
+    env DCONF_PROFILE=ibus dconf write /desktop/ibus/general/preload-engines "['BambooUs', 'Bamboo']" && gsettings set org.gnome.desktop.input-sources sources "[('xkb', 'us'), ('ibus', 'Bamboo')]"
+else
+    printf "${OK} ${MAGENTA}ibus-bamboo${RESET} is already installed. Skipping\n"
+fi
 printf "\n%.0s" {1..1}
 
 
@@ -244,35 +267,36 @@ printf "\n%.0s" {1..1}
 
 # ----------install yazi----------
 printf "${NOTE} Installing ${SKY_BLUE}yazi${RESET} ...${RESET}\n"
-yazi=(
-    ffmpeg 
-    7zip 
-    jq 
-    poppler-utils 
-    fd-find 
-    ripgrep 
-    fzf 
-    zoxide 
-    imagemagick
-)
-for i in "${yazi[@]}"; do
-    install_package "$i"
-done
-
 if command -v yazi >/dev/null 2>&1; then
     printf "${OK} ${MAGENTA}yazi${RESET} is already installed. Skipping\n"
 else
+    yazi=(
+        ffmpeg 
+        7zip 
+        jq 
+        poppler-utils 
+        fd-find 
+        ripgrep 
+        fzf 
+        zoxide 
+        imagemagick
+    )
+    for i in "${yazi[@]}"; do
+        install_package "$i"
+    done
     stdbuf -oL sudo snap install yazi --classic
     # double check if the package was re-installed successfully
     if command -v yazi >/dev/null 2>&1; then
-        printf "${OK} ${YELLOW}yazi${RESET} has been installed successfully\n"
+        printf "${OK} ${SKY_BLUE}yazi${RESET} has been installed successfully\n"
     else
         printf "${ERROR} ${YELLOW}yazi${RESET} failed to install.\n"
     fi
 fi
 # copy the yazi configuration file
 printf "${NOTE} Copying ${SKY_BLUE}yazi configuration${RESET} ...${RESET}\n"
-mkdir -p ~/.config/yazi
+if [ ! -d "$HOME/.config/yazi" ]; then
+    mkdir -p "$HOME/.config/yazi"
+fi
 cp -r yazi/* ~/.config/yazi
 printf "\n%.0s" {1..1}
 
@@ -293,11 +317,12 @@ else
     sudo ./install.sh
     rm -rf ../anki $ANKI_FILE
     if command -v anki > /dev/null 2>&1; then
-        printf "${OK} ${YELLOW}Anki${RESET} has been installed successfully\n"
+        printf "${OK} ${SKY_BLUE}Anki${RESET} has been installed successfully\n"
     else
         printf "${ERROR} ${YELLOW}Anki${RESET} failed to install.\n"
     fi
 fi
+printf "\n%.0s" {1..1}
 
 
 # install tailscale
@@ -310,11 +335,12 @@ else
     apt_update
     install_package "tailscale"
     if command -v tailscale > /dev/null 2>&1; then
-        printf "${OK} ${YELLOW}Tailscale${RESET} has been installed successfully\n"
+        printf "${OK} ${SKY_BLUE}Tailscale${RESET} has been installed successfully\n"
     else
         printf "${ERROR} ${YELLOW}Tailscale${RESET} failed to install.\n"
     fi
 fi
+printf "\n%.0s" {1..1}
 
 
 # ----------install yt-dlp----------
@@ -326,8 +352,9 @@ else
     apt_update
     install_package "yt-dlp"
     if command -v yt-dlp > /dev/null 2>&1; then
-        printf "${OK} ${YELLOW}yt-dlp${RESET} has been installed successfully\n"
+        printf "${OK} ${SKY_BLUE}yt-dlp${RESET} has been installed successfully\n"
     else
         printf "${ERROR} ${YELLOW}yt-dlp${RESET} failed to install.\n"
     fi
 fi
+printf "\n%.0s" {1..1}
